Convert to .bp, cleanup
Drop explicit `std=` versions.

All the conditional logic like `TARGET_DISABLE_RMNT` is
Qcom-specific and, to be honest, total crap. We don't need
it.

Compilation with kernel headers isn't really used as well.
Would be nice, but (almost) nobody does it. Also, there
isn't a nice way to do this in blueprints, save for defining
soong vars and writing plugins in go for _every_ module.
Yikes.

Use generated kernel headers instead - for now.
Note: This is actually needed and not optional, because
it seems kernel headers are no longer implicitly included
when using blueprints.

LOCAL_CLANG is obsolete since clang is the default compiler
now.

We don't need any build guards for now, but could put this
behind a soong namespace if necessary.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>
Signed-off-by: ix5 <ix5@users.noreply.github.com>